requirements.txt
numpy
pandas
tensorflow
keras
opencv-python
matplotlib
Pillow
scikit-learn
